Fixed the popular tracker bug
The main problem we ran into when dealing with the popular tracker was that
the RPC that the popular tracker returns is TOO LARGE to fit in the tracker's
buffer completely. Therefore, originally, the program was never able to read
to the end of the RPC on one pass, so read_tracker_response() would "end
prematurely".

I fixed this by having the caller of read_tracker_response() allocate a larger
array in which multiple calls to read_tracker_response() would append its data
to untill read_tracker_response() reaches the end of the RPC.

However, there was a secondary issue that came up, which I hadn't noticed
earlier. Because the entire program shared only a single tracker (even though
we did a memcpy on tracker_task...we only called start_tracker once), the
different threads that call read_tracker_response() would often read each
others' responses! This is because read_tracker_response is no longer atomic.
One solution was to obviously put a mutex on the read_tracker_response()
section. The other solution was to simple open a new tracker connection for
each download.

I have implemented the latter fix. I want to play around with the mutex fix
now. Expect more optimizations to the popular-tracker bug fix.