Fixed the racing issue with read_tracker_response().
IM SO HAPPY.

If we only connect to a single tracker, we are able to have each download
thread issue a HAVE *file* command after each successful download and have
the tracker correctly track those files back to us.

If we connect to multiple trackers (to avoid these racing issues), we would
have difficulties telling the tracker that each file we download belongs to
the same peer (us).

The problem with using a single tracker is that we need to make sure that only
a single thread is allowed to talk to the tracker at one time. Therefore we
must use mutexes in sections of the code in which multiple threads CAN talk to
the tracker at the same time. In our case, it would have been the main thread
issuing a WANT command at the same time a download thread is issuing a HAVE
command.